pyproject.toml
[project]
name = "smartnewsanalyticsplatform"
version = "0.1.0"
description = "Smart News Analytics Platform with PostgreSQL advanced features"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.0",
    "djangorestframework>=3.15.0",
    "psycopg[binary]>=3.2.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "sentence-transformers>=3.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "django-cors-headers>=4.3.0",
    "django-ratelimit>=4.1.0",
    "channels>=4.0.0",
    "channels-redis>=4.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "SmartNewsAnalyticsPlatform"]